requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
streamlit==1.38.0
requests==2.32.3
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1
langchain-core==0.3.6
faiss-cpu==1.13.2
pymupdf==1.25.3
pdfplumber==0.11.4
openai==1.47.0
tiktoken==0.7.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
loguru==0.7.2
